+++ sdk/nodejs/src/index.ts
@@ -1,2 +1,3 @@
 export { XR2Client } from "./client";
+export { renderPrompt } from "./render";
 export * from "./types";